refactor: LottoService를 LottoMachineService와 LottoResultService로 분리
